requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
anyio=4.2.0=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
babel=2.11.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
cftime=1.6.3=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=py39h06a4308_0
contourpy=1.2.0=pypi_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.3.101=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.3.101=0
cuda-nsight-compute=12.3.2=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.3.101=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.3.101=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.3.101=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.3.101=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.49.0=pypi_0
freetype=2.12.1=h4a9f257_0
fvcore=0.1.5.post20221221=pyhd8ed1ab_0
gds-tools=1.8.1.2=0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.10.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
imageio=2.34.0=pypi_0
importlib-metadata=7.0.1=py39h06a4308_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.9=py39
ipykernel=6.28.0=py39h06a4308_0
ipython=8.15.0=py39h06a4308_0
ipywidgets=8.0.4=py39h06a4308_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.3=py39h06a4308_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py39h06a4308_0
jsonschema-specifications=2023.7.1=py39h06a4308_0
jupyter=1.0.0=py39h06a4308_9
jupyter-lsp=2.2.0=py39h06a4308_0
jupyter_client=8.6.0=py39h06a4308_0
jupyter_console=6.6.3=py39h06a4308_0
jupyter_core=5.5.0=py39h06a4308_0
jupyter_events=0.8.0=py39h06a4308_0
jupyter_server=2.10.0=py39h06a4308_0
jupyter_server_terminals=0.4.4=py39h06a4308_1
jupyterlab=4.0.11=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py39h06a4308_0
jupyterlab_widgets=3.0.9=py39h06a4308_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.8.1.2=0
libcufile-dev=1.8.1.2=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libglib=2.78.4=hdc74915_0
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=py39h06a4308_0
mdurl=0.1.2=pypi_0
meshio=5.3.5=pypi_0
mistune=2.0.4=py39h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mysql=5.7.24=h721c034_2
nbclient=0.8.0=py39h06a4308_0
nbconvert=7.10.0=py39h06a4308_0
nbformat=5.9.2=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
netcdf4=1.6.5=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
notebook=7.0.8=py39h06a4308_0
notebook-shim=0.2.3=py39h06a4308_0
nsight-compute=2023.3.1.1=0
numpy=1.26.3=py39h5f9d8c6_0
numpy-base=1.26.3=py39hb5e798b_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_0
overrides=7.4.0=py39h06a4308_0
packaging=23.1=py39h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=hebb0a14_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.2.0=py39h5eee18b_0
pip=23.3.1=py39h06a4308_0
platformdirs=3.10.0=py39h06a4308_0
plotly=5.19.0=pypi_0
ply=3.11=py39h06a4308_0
plyfile=1.0.3=pypi_0
point-cloud-utils=0.30.4=pypi_0
pooch=1.8.0=pypi_0
portalocker=2.8.2=py39hf3d152e_1
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.43=py39h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py39h06a4308_1
pyparsing=3.1.1=pypi_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python-json-logger=2.0.7=py39h06a4308_0
python_abi=3.9=2_cp39
pytorch=1.13.0=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytorch3d=0.7.5=py39_cu116_pyt1130
pytz=2023.3.post1=py39h06a4308_0
pyvista=0.43.3=pypi_0
pywavefront=1.3.3=pypi_0
pyyaml=6.0.1=py39hd1e30aa_1
pyzmq=25.1.2=py39h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
qtconsole=5.5.0=py39h06a4308_0
qtpy=2.4.1=py39h06a4308_0
readline=8.2=h5eee18b_0
referencing=0.30.2=py39h06a4308_0
requests=2.31.0=py39h06a4308_1
rfc3339-validator=0.1.4=py39h06a4308_0
rfc3986-validator=0.1.1=py39h06a4308_0
rich=13.7.0=pypi_0
rpds-py=0.10.6=py39hb02cf49_0
scikit-image=0.22.0=pypi_0
scipy=1.12.0=pypi_0
scooby=0.9.2=pypi_0
send2trash=1.8.2=py39h06a4308_0
setuptools=68.2.2=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39h06a4308_0
soupsieve=2.5=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.9.0=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
termcolor=2.4.0=pyhd8ed1ab_0
terminado=0.17.1=py39h06a4308_0
tifffile=2024.2.12=pypi_0
tinycss2=1.2.1=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py39h06a4308_0
torchvision=0.14.0=py39_cu116
tornado=6.3.3=py39h5eee18b_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.7.1=py39h06a4308_0
trimesh=4.1.4=pypi_0
typing-extensions=4.9.0=py39h06a4308_1
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2023d=h04d1e81_0
urllib3=2.1.0=py39h06a4308_1
vtk=9.3.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.5=py39h06a4308_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
